move programmatic datalayer code to tx-map
