+++ src/interface/src/app/treatments/treatment-map/treatment-map.component.html
@@ -55,9 +55,6 @@
     [layout]="{ visibility: 'none' }">
   </mgl-layer>
 
-  <app-map-data-layer [mapLibreMap]="mapLibreMap"
-    *appFeatureFlag="'statewide_datalayers'"></app-map-data-layer>
-
   <app-planning-area-layer> </app-planning-area-layer>
 
   <app-map-project-areas
